requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0
pymysql>=1.1.0
tushare>=1.4.0
openai>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
xgboost>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
